internalize into httpclient rate limiter
