+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/sink/client/AbstractHttpClient.java
@@ -1,5 +1,8 @@
 package io.github.clescot.kafka.connect.http.sink.client;
 
+import dev.failsafe.RateLimiter;
+import io.github.clescot.kafka.connect.http.core.HttpExchange;
+
 import javax.annotation.Nullable;
 import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.SSLContext;
@@ -14,6 +17,7 @@ import java.security.*;
 import java.security.cert.CertificateException;
 import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
 
 import static io.github.clescot.kafka.connect.http.sink.HttpSinkConfigDefinition.*;
 
@@ -21,6 +25,7 @@ public abstract class AbstractHttpClient<Req, Res> implements HttpClient<Req, Re
 
     private static final String DEFAULT_SSL_PROTOCOL = "SSL";
     protected Map<String, String> config;
+    private Optional<RateLimiter<HttpExchange>> rateLimiter = Optional.empty();
 
     public AbstractHttpClient(Map<String, String> config) {
         this.config = config;
@@ -98,6 +103,31 @@ public abstract class AbstractHttpClient<Req, Res> implements HttpClient<Req, Re
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    @Override
+    public CompletableFuture<Res> call(Req request){
+        try {
+            Optional<RateLimiter<HttpExchange>> rateLimiter = getRateLimiter();
+            if (rateLimiter.isPresent()) {
+                rateLimiter.get().acquirePermits(HttpClient.ONE_HTTP_REQUEST);
+                LOGGER.debug("permits acquired request:'{}'", request);
+            }
+            return nativeCall(request);
+        } catch (InterruptedException e) {
+            LOGGER.error("Failed to acquire execution permit from the rate limiter {} ", e.getMessage());
+            throw new HttpException(e.getMessage());
+        }
+    }
+
+    @Override
+    public void setRateLimiter(RateLimiter<HttpExchange> rateLimiter) {
+        this.rateLimiter = Optional.ofNullable(rateLimiter);
+    }
+
+    @Override
+    public Optional<RateLimiter<HttpExchange>> getRateLimiter() {
+        return this.rateLimiter;
     }
 
 }
